Fix the build: update deps and add a nil check
Solution similar to https://github.com/kozalosev/PostSuggesterBot/pull/3

+++ db/repo/common_test.go
@@ -74,7 +74,9 @@ func setup() {
 }
 
 func shutDown() {
-	if err := container.Terminate(ctx); err != nil {
-		panic(fmt.Sprintf("failed to terminate container: %s", err.Error()))
+	if container != nil {
+		if err := container.Terminate(ctx); err != nil {
+			panic(fmt.Sprintf("failed to terminate container: %s", err.Error()))
+		}
 	}
 }
